Move/Rename JS Regex to Bridge.Regex; Adjust namespaces;
