+++ Bridge/Bridge/Linq/Enumerable.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text.RegularExpressions;
+using Bridge.Text.RegularExpressions;
 
 namespace Bridge.Linq
 {
