+++ Bridge/Bridge/Text/RegularExpressions/RegexMatch.cs
@@ -1,6 +1,6 @@
 using Bridge;
 
-namespace System.Text.RegularExpressions
+namespace Bridge.Text.RegularExpressions
 {
     [External]
     [Name("Object")]
