+++ Bridge/System/Text/RegularExpressions/RegexMatch.cs
@@ -1,6 +1,6 @@
 ﻿using Bridge;
 
-namespace System.Text.RegularExpressions.CoreFx
+namespace System.Text.RegularExpressions
 {
     /// <summary>
     /// Represents the results from a single regular expression match.
