+++ Tests/Batch1/Constants.cs
@@ -38,6 +38,7 @@ namespace Bridge.ClientTest
         public const string MODULE_NULLABLE = "Nullable";
         public const string MODULE_STRING = "String";// "String, StringBuilder"
         public const string MODULE_REGEX = "Regex";
+        public const string MODULE_REGEX_JS = "RegexJS";
         public const string MODULE_ENUM = "Enum";
 
         public const string MODULE_MATH = "Math";
