+++ Tests/Batch1/Text/RegularExpressions/Msdn/RegexEscapeTests.cs
@@ -1,9 +1,8 @@
 ﻿using System.Collections.Generic;
-using System.Text.RegularExpressions.CoreFx;
-using Bridge.Html5;
+using System.Text.RegularExpressions;
 using Bridge.Test;
 
-namespace Bridge.ClientTest.Text.RegularExpressions.CoreFx
+namespace Bridge.ClientTest.Text.RegularExpressions.Msdn
 {
     [Category(Constants.MODULE_REGEX)]
     [TestFixture(TestNameFormat = "Regex.Escape - {0}")]
