+++ Tests/Batch1/Text/RegularExpressions/Msdn/RegexReplaceTests.cs
@@ -1,8 +1,8 @@
 ﻿using System;
-using System.Text.RegularExpressions.CoreFx;
+using System.Text.RegularExpressions;
 using Bridge.Test;
 
-namespace Bridge.ClientTest.Text.RegularExpressions.CoreFx
+namespace Bridge.ClientTest.Text.RegularExpressions.Msdn
 {
     [Category(Constants.MODULE_REGEX)]
     [TestFixture(TestNameFormat = "Regex.Replace - {0}")]
